backend/requirements.txt
tensorflow>=2.16.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
keras>=3.0.0
Pillow>=10.0.0
jupyter>=1.0.0
tensorflowjs
imagehash>=4.3.1
scipy
seaborn
tqdm
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart
